Adding new distribution functions for tool and virus based on entity (#133)
* Adding new distribution functions for tool and virus based on entity

* Running the single header

* Bumping to 11

* Apply suggestions from code review


* Addressing review comments

* Addressing more comments

* Fix CSV output: quote state labels for data.table compatibility (#136)

* Initial plan

* Add test for virus history recording


* Fix CSV output: add quotes to state labels in virus_hist for proper data.table parsing


* Add comprehensive CSV format validation tests and fix header spacing


* Remove hard-coded paths from tests for better portability


* Make CSV format tests more precise


* Renaming files and adding them to the main tests

* Adding another missing test and renaming

* Fix test file paths to use save directories per Makefile convention


---------


* Adding the changes to the single header

---------

+++ tests/06-mixing.cpp
@@ -52,7 +52,7 @@ EPIWORLD_TEST_CASE("SIRMixing", "[SIR-mixing]") {
 
     for (const auto & a : model.get_agents())
     {
-        if (a.get_state() != model.state_of("Susceptible"))
+        if (static_cast<int>(a.get_state()) != model.state_of("Susceptible"))
         {
             if (a.get_entity(0).get_id() == 0)
             {
@@ -92,7 +92,7 @@ EPIWORLD_TEST_CASE("SIRMixing", "[SIR-mixing]") {
         {
             n_right++;
         }
-        else if (a.get_state() != model.state_of("Susceptible"))
+        else if (static_cast<int>(a.get_state()) != model.state_of("Susceptible"))
         {
             if (a.get_entity(0).get_id() == 1)
             {
